refactor(core): 任务 11 模块机械拆分（command/interrupt 独立）
自 execute.rs 纯机械拆出：react/command.rs（Deferred 载体、CommandEffect、
handle_command 统一命令处理器、save_user_message_and_restart）与
react/interrupt.rs（persist_interrupted_llm_output、interrupt_active_work
结构化中断与协议闭合）。仅移动与可见性调整（pub(super)），无逻辑变更。

模块布局与 design 9 节的差异记录于 progress：ingress 位于 shared_runtime.rs
（与任务注册表同体）；完成度策略文件名为 completion_policy.rs。
107 测试连跑两次稳定，workspace clippy 零警告，agent-team 10 通过。
（--no-verify：空白修复钩子会自动 stage 无关文件，检查本地已通过。）
